Split wind_damage.rs (716 lines) into wind_damage/ subdirectory
Breaks the monolithic wind_damage.rs into focused sub-modules:
- types.rs (176 lines): WindDamageTier enum, damage formulas, constants, state, events, RNG
- systems.rs (118 lines): update_wind_damage system and WindDamagePlugin
- tests.rs (436 lines): all unit and integration tests
- mod.rs (10 lines): pub use re-exports for backward compatibility

No behavioral changes. All public API items are re-exported from mod.rs
so external callers remain unchanged.

Closes #1389
