+++ crates/simulation/src/wind_damage/tests.rs
@@ -1,280 +1,15 @@
-use bevy::prelude::*;
-use serde::{Deserialize, Serialize};
-// Storm wind damage thresholds (Beaufort scale)
-use crate::config::{GRID_HEIGHT, GRID_WIDTH};
-use crate::trees::TreeGrid;
-use crate::wind::WindState;
-use crate::TickCounter;
-
-// =============================================================================
-// Wind Damage Tiers (Beaufort-inspired)
-// =============================================================================
-
-/// Beaufort-inspired wind damage classification based on normalized wind speed [0, 1].
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize, Default)]
-pub enum WindDamageTier {
-    /// Speed 0.0 - 0.15: No damage.
-    #[default]
-    Calm,
-    /// Speed 0.15 - 0.3: No damage, light wind.
-    Breezy,
-    /// Speed 0.3 - 0.45: Minor effects, no structural damage.
-    Strong,
-    /// Speed 0.45 - 0.6: Light structural risk begins.
-    Gale,
-    /// Speed 0.6 - 0.75: Moderate damage, power lines at risk.
-    Storm,
-    /// Speed 0.75 - 0.9: Significant damage, trees knocked down.
-    Severe,
-    /// Speed 0.9 - 0.95: Extreme damage to structures.
-    HurricaneForce,
-    /// Speed > 0.95: Catastrophic damage.
-    Extreme,
-}
-
-impl WindDamageTier {
-    /// Classify a normalized wind speed [0, 1] into a damage tier.
-    pub fn from_speed(speed: f32) -> Self {
-        if speed < 0.15 {
-            WindDamageTier::Calm
-        } else if speed < 0.3 {
-            WindDamageTier::Breezy
-        } else if speed < 0.45 {
-            WindDamageTier::Strong
-        } else if speed < 0.6 {
-            WindDamageTier::Gale
-        } else if speed < 0.75 {
-            WindDamageTier::Storm
-        } else if speed < 0.9 {
-            WindDamageTier::Severe
-        } else if speed < 0.95 {
-            WindDamageTier::HurricaneForce
-        } else {
-            WindDamageTier::Extreme
-        }
-    }
-
-    /// Human-readable label for UI display.
-    pub fn label(self) -> &'static str {
-        match self {
-            WindDamageTier::Calm => "Calm",
-            WindDamageTier::Breezy => "Breezy",
-            WindDamageTier::Strong => "Strong",
-            WindDamageTier::Gale => "Gale",
-            WindDamageTier::Storm => "Storm",
-            WindDamageTier::Severe => "Severe",
-            WindDamageTier::HurricaneForce => "Hurricane Force",
-            WindDamageTier::Extreme => "Extreme",
-        }
-    }
-}
-
-// =============================================================================
-// Damage formulas
-// =============================================================================
-
-/// Wind damage threshold: damage begins above this normalized speed.
-const WIND_DAMAGE_THRESHOLD: f32 = 0.4;
-
-/// Power outage threshold: outage probability begins above this speed.
-const POWER_OUTAGE_THRESHOLD: f32 = 0.6;
-
-/// Tree knockdown threshold: tree damage begins above this speed.
-const TREE_KNOCKDOWN_THRESHOLD: f32 = 0.6;
-
-/// Calculate wind damage amount using cubic formula.
-/// Returns 0.0 for speeds <= 0.4, otherwise `(speed - 0.4)^3 * 1000`.
-pub fn wind_damage_amount(speed: f32) -> f32 {
-    if speed <= WIND_DAMAGE_THRESHOLD {
-        return 0.0;
-    }
-    let excess = speed - WIND_DAMAGE_THRESHOLD;
-    excess * excess * excess * 1000.0
-}
-
-/// Calculate power outage probability based on wind speed.
-/// Returns 0.0 for speeds <= 0.6, scaling up to ~1.0 at extreme speeds.
-/// Formula: `((speed - 0.6) / 0.4)^2` clamped to [0, 1].
-pub fn power_outage_probability(speed: f32) -> f32 {
-    if speed <= POWER_OUTAGE_THRESHOLD {
-        return 0.0;
-    }
-    let factor = (speed - POWER_OUTAGE_THRESHOLD) / 0.4;
-    (factor * factor).min(1.0)
-}
-
-/// Calculate tree knockdown probability based on wind speed.
-/// Returns 0.0 for speeds <= 0.6, scaling up for higher speeds.
-/// Formula: `((speed - 0.6) / 0.4)^2 * 0.1` per tree per update tick.
-pub fn tree_knockdown_probability(speed: f32) -> f32 {
-    if speed <= TREE_KNOCKDOWN_THRESHOLD {
-        return 0.0;
-    }
-    let factor = (speed - TREE_KNOCKDOWN_THRESHOLD) / 0.4;
-    (factor * factor * 0.1).min(1.0)
-}
-
-// =============================================================================
-// Wind Damage State (resource)
-// =============================================================================
-
-/// Resource tracking accumulated wind damage during a storm.
-#[derive(Resource, Debug, Clone, Serialize, Deserialize)]
-pub struct WindDamageState {
-    /// Current wind damage tier classification.
-    #[serde(default)]
-    pub current_tier: WindDamageTier,
-    /// Accumulated building damage this storm (cumulative damage points).
-    #[serde(default)]
-    pub accumulated_building_damage: f32,
-    /// Number of trees knocked down during this storm.
-    #[serde(default)]
-    pub trees_knocked_down: u32,
-    /// Whether a power outage is currently active due to wind.
-    #[serde(default)]
-    pub power_outage_active: bool,
-}
-
-impl Default for WindDamageState {
-    fn default() -> Self {
-        Self {
-            current_tier: WindDamageTier::Calm,
-            accumulated_building_damage: 0.0,
-            trees_knocked_down: 0,
-            power_outage_active: false,
-        }
-    }
-}
-
-// =============================================================================
-// Wind Damage Event
-// =============================================================================
-
-/// Event fired when wind damage occurs, for notification to other systems.
-#[derive(Event, Debug, Clone)]
-pub struct WindDamageEvent {
-    /// The damage tier that triggered this event.
-    pub tier: WindDamageTier,
-    /// Amount of building damage dealt this tick.
-    pub building_damage: f32,
-    /// Number of trees knocked down this tick.
-    pub trees_knocked: u32,
-    /// Whether power outage was triggered.
-    pub power_outage: bool,
-}
-
-// =============================================================================
-// Deterministic pseudo-random (splitmix64, matching wind.rs pattern)
-// =============================================================================
-
-fn splitmix64(mut x: u64) -> u64 {
-    x = x.wrapping_add(0x9e3779b97f4a7c15);
-    x = (x ^ (x >> 30)).wrapping_mul(0xbf58476d1ce4e5b9);
-    x = (x ^ (x >> 27)).wrapping_mul(0x94d049bb133111eb);
-    x ^ (x >> 31)
-}
-
-/// Returns a deterministic pseudo-random f32 in [0.0, 1.0) based on seed.
-fn rand_f32(seed: u64) -> f32 {
-    let hash = splitmix64(seed);
-    (hash % 1_000_000) as f32 / 1_000_000.0
-}
-
-// =============================================================================
-// Systems
-// =============================================================================
-
-/// Wind damage update interval in ticks (aligns with wind update interval).
-const WIND_DAMAGE_INTERVAL: u64 = 100;
-
-/// Updates wind damage state each interval based on current wind speed.
-///
-/// - Classifies wind into a damage tier
-/// - Accumulates building damage for speeds > 0.4
-/// - Probabilistically knocks down trees at high wind speeds
-/// - Sets power outage flag based on outage probability
-///
-/// Resets accumulated counters when wind drops below damage threshold.
-pub fn update_wind_damage(
-    tick: Res<TickCounter>,
-    wind: Res<WindState>,
-    mut state: ResMut<WindDamageState>,
-    mut tree_grid: ResMut<TreeGrid>,
-    mut events: EventWriter<WindDamageEvent>,
-) {
-    if tick.0 == 0 || !tick.0.is_multiple_of(WIND_DAMAGE_INTERVAL) {
-        return;
-    }
-
-    let speed = wind.speed;
-    let tier = WindDamageTier::from_speed(speed);
-    state.current_tier = tier;
-
-    // If below damage threshold, reset storm counters and exit
-    if speed <= WIND_DAMAGE_THRESHOLD {
-        // Only reset when transitioning from a damaging state
-        if state.accumulated_building_damage > 0.0 || state.trees_knocked_down > 0 {
-            state.accumulated_building_damage = 0.0;
-            state.trees_knocked_down = 0;
-        }
-        state.power_outage_active = false;
-        return;
-    }
-
-    // --- Building damage ---
-    let damage = wind_damage_amount(speed);
-    state.accumulated_building_damage += damage;
-
-    // --- Power outage ---
-    let outage_prob = power_outage_probability(speed);
-    let outage_seed = tick.0.wrapping_mul(0xdeadbeef_cafebabe);
-    let outage_roll = rand_f32(outage_seed);
-    state.power_outage_active = outage_roll < outage_prob;
-
-    // --- Tree knockdown ---
-    let knockdown_prob = tree_knockdown_probability(speed);
-    let mut trees_knocked_this_tick: u32 = 0;
-
-    if knockdown_prob > 0.0 {
-        // Iterate over the grid to find trees and probabilistically knock them down.
-        // Use deterministic hash based on tick + position for each cell.
-        for y in 0..GRID_HEIGHT {
-            for x in 0..GRID_WIDTH {
-                if tree_grid.has_tree(x, y) {
-                    let cell_seed = tick
-                        .0
-                        .wrapping_mul(0x517cc1b727220a95)
-                        .wrapping_add((y * GRID_WIDTH + x) as u64);
-                    let roll = rand_f32(cell_seed);
-                    if roll < knockdown_prob {
-                        tree_grid.set(x, y, false);
-                        trees_knocked_this_tick += 1;
-                    }
-                }
-            }
-        }
-    }
-
-    state.trees_knocked_down += trees_knocked_this_tick;
-
-    // Fire event if any damage occurred
-    if damage > 0.0 || trees_knocked_this_tick > 0 || state.power_outage_active {
-        events.send(WindDamageEvent {
-            tier,
-            building_damage: damage,
-            trees_knocked: trees_knocked_this_tick,
-            power_outage: state.power_outage_active,
-        });
-    }
-}
-
-// =============================================================================
-// Tests
-// =============================================================================
-
 #[cfg(test)]
 mod tests {
-    use super::*;
+    use bevy::prelude::*;
+
+    use crate::trees::TreeGrid;
+    use crate::wind::WindState;
+    use crate::wind_damage::systems::update_wind_damage;
+    use crate::wind_damage::types::{
+        power_outage_probability, rand_f32, splitmix64, tree_knockdown_probability,
+        wind_damage_amount, WindDamageEvent, WindDamageState, WindDamageTier,
+    };
+    use crate::TickCounter;
 
     // -----------------------------------------------------------------------
     // Unit tests for damage tier classification
@@ -697,20 +432,5 @@ mod tests {
                 val
             );
         }
-    }
-}
-
-pub struct WindDamagePlugin;
-
-impl Plugin for WindDamagePlugin {
-    fn build(&self, app: &mut App) {
-        app.init_resource::<WindDamageState>()
-            .add_event::<WindDamageEvent>()
-            .add_systems(
-                FixedUpdate,
-                update_wind_damage
-                    .after(crate::imports_exports::process_trade)
-                    .in_set(crate::SimulationSet::Simulation),
-            );
     }
 }
